Reorganized the mess
