+++ SteelSeriesAPI/Sonar/Events/SonarAudienceMonitoringEvent.cs
@@ -1,4 +1,4 @@
-namespace SteelSeriesAPI.Events;
+namespace SteelSeriesAPI.Sonar.Events;
 
 public class SonarAudienceMonitoringEvent : EventArgs
 {
