+++ SteelSeriesAPI/Sonar/Events/SonarChatMixEvent.cs
@@ -1,4 +1,4 @@
-namespace SteelSeriesAPI.Events;
+namespace SteelSeriesAPI.Sonar.Events;
 
 public class SonarChatMixEvent : EventArgs
 {
