+++ SteelSeriesAPI/Sonar/Http/SonarHttpProvider.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections;
 using System.Text.Json;
-using SteelSeriesAPI.Interfaces;
+using SteelSeriesAPI.Sonar.Interfaces;
 using SteelSeriesAPI.Sonar.Enums;
 using SteelSeriesAPI.Sonar.Models;
