+++ SteelSeriesAPI/Sonar/SonarBridge.cs
@@ -1,11 +1,12 @@
 ﻿using System.Security.Principal;
 using SteelSeriesAPI.Interfaces;
-using SteelSeriesAPI.Sonar;
 using SteelSeriesAPI.Sonar.Enums;
 using SteelSeriesAPI.Sonar.Http;
+using SteelSeriesAPI.Sonar.Interfaces;
+using SteelSeriesAPI.Sonar.Managers;
 using SteelSeriesAPI.Sonar.Models;
 
-namespace SteelSeriesAPI;
+namespace SteelSeriesAPI.Sonar;
 
 /// <summary>
 /// The Sonar object, to control Sonar<br/>Allow you to listen for event, get or set volumes, muted states, ...
